Renaming the repo for release.
