+++ database_factory/cloud/aws/auth.py
