+++ database_factory/cloud/gcp/resource_manager.py
